Rename DisplayList::DisplayListItem to DisplayList::Item and move it to a separate file https://bugs.webkit.org/show_bug.cgi?id=260668 rdar://114391530
Reviewed by Kimmo Kinnunen.

Remove the redundancy in the name and create dedicated source and header files
for the DisplayList::Item. Move its existing functions from DisplayListItems.cpp
and DisplayListReplayer.cpp to the new files.

* Source/WebCore/Headers.cmake:
* Source/WebCore/Sources.txt:
* Source/WebCore/WebCore.xcodeproj/project.pbxproj:
* Source/WebCore/platform/graphics/displaylists/DisplayList.cpp:
(WebCore::DisplayList::DisplayList::append):
(WebCore::DisplayList::DisplayList::asText const):
(WebCore::DisplayList::DisplayList::dump const):
* Source/WebCore/platform/graphics/displaylists/DisplayList.h:
(WebCore::DisplayList::DisplayList::items const):
(WebCore::DisplayList::DisplayList::items):
* Source/WebCore/platform/graphics/displaylists/DisplayListItem.cpp: Copied from Source/WebCore/platform/graphics/displaylists/DisplayListReplayer.cpp.
(WebCore::DisplayList::isValid):
(WebCore::DisplayList::applyImageBufferItem):
(WebCore::DisplayList::applyNativeImageItem):
(WebCore::DisplayList::applySourceImageItem):
(WebCore::DisplayList::applySetStateItem):
(WebCore::DisplayList::applyDrawGlyphs):
(WebCore::DisplayList::applyDrawDecomposedGlyphs):
(WebCore::DisplayList::applyItem):
(WebCore::DisplayList::shouldDumpItem):
(WebCore::DisplayList::dumpItem):
(WebCore::DisplayList::operator<<):
* Source/WebCore/platform/graphics/displaylists/DisplayListItem.h: Added.
* Source/WebCore/platform/graphics/displaylists/DisplayListItems.cpp:
(WebCore::DisplayList::shouldDumpDisplayListItem): Deleted.
(WebCore::DisplayList::dumpDisplayListItem): Deleted.
* Source/WebCore/platform/graphics/displaylists/DisplayListItems.h:
(WebCore::DisplayList::isValid): Deleted.
(WebCore::DisplayList::operator<<): Deleted.
* Source/WebCore/platform/graphics/displaylists/DisplayListRecorderImpl.h:
(WebCore::DisplayList::RecorderImpl::append):
* Source/WebCore/platform/graphics/displaylists/DisplayListReplayer.cpp:
(WebCore::DisplayList::Replayer::Replayer):
(WebCore::DisplayList::Replayer::replay):
(WebCore::DisplayList::applyImageBufferItem): Deleted.
(WebCore::DisplayList::applyNativeImageItem): Deleted.
(WebCore::DisplayList::applySourceImageItem): Deleted.
(WebCore::DisplayList::applySetStateItem): Deleted.
(WebCore::DisplayList::applyDrawGlyphs): Deleted.
(WebCore::DisplayList::applyDrawDecomposedGlyphs): Deleted.
(WebCore::DisplayList::Replayer::applyItem): Deleted.
* Source/WebCore/platform/graphics/displaylists/DisplayListReplayer.h:
* Tools/TestWebKitAPI/Tests/WebCore/cg/DisplayListTestsCG.cpp:
(TestWebKitAPI::TEST):

Canonical link: https://commits.webkit.org/267431@main
